+++ app/src/main/java/org/stepik/android/view/injection/profile/ProfileCoursesComponent.kt
@@ -1,0 +1,14 @@
+package org.stepik.android.view.injection.profile
+
+import dagger.Subcomponent
+import org.stepik.android.view.profile_courses.ui.fragment.ProfileCoursesFragment
+
+@ProfileCoursesScope
+@Subcomponent(modules = [ProfileCoursesPresentationModule::class])
+interface ProfileCoursesComponent {
+    @Subcomponent.Builder
+    interface Builder {
+        fun build(): ProfileCoursesComponent
+    }
+    fun inject(profileCoursesFragment: ProfileCoursesFragment)
+}
