+++ app/src/main/java/org/stepik/android/view/injection/profile/ProfileCoursesScope.kt
@@ -1,0 +1,7 @@
+package org.stepik.android.view.injection.profile
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ProfileCoursesScope
