+++ app/src/main/java/org/stepik/android/view/injection/profile/ProfileModule.kt
@@ -11,20 +11,14 @@ import io.reactivex.subjects.BehaviorSubject
 import org.stepic.droid.di.qualifiers.BackgroundScheduler
 import org.stepik.android.domain.profile.model.ProfileData
 import org.stepik.android.presentation.base.injection.ViewModelKey
-import org.stepik.android.presentation.course_continue.CourseContinueView
 import org.stepik.android.presentation.profile.ProfilePresenter
 import org.stepik.android.presentation.profile_achievements.ProfileAchievementsPresenter
 import org.stepik.android.presentation.profile_activities.ProfileActivitiesPresenter
 import org.stepik.android.presentation.profile_certificates.ProfileCertificatesPresenter
-import org.stepik.android.presentation.profile_courses.ProfileCoursesPresenter
-import org.stepik.android.presentation.profile_courses.ProfileCoursesView
 import org.stepik.android.presentation.profile_detail.ProfileDetailPresenter
 import org.stepik.android.presentation.profile_id.ProfileIdPresenter
 import org.stepik.android.presentation.profile_links.ProfileLinksPresenter
 import org.stepik.android.presentation.profile_notification.ProfileNotificationPresenter
-import ru.nobird.android.presentation.base.DefaultPresenterViewContainer
-import ru.nobird.android.presentation.base.PresenterViewContainer
-import ru.nobird.android.presentation.base.ViewContainer
 
 @Module
 abstract class ProfileModule {
@@ -68,16 +62,8 @@ abstract class ProfileModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(ProfileCoursesPresenter::class)
-    internal abstract fun bindProfileCoursesPresenter(ProfileCoursesPresenter: ProfileCoursesPresenter): ViewModel
-
-    @Binds
-    @IntoMap
     @ViewModelKey(ProfileCertificatesPresenter::class)
     internal abstract fun bindProfileCertificatePresenter(profileCertificatesPresenter: ProfileCertificatesPresenter): ViewModel
-
-    @Binds
-    internal abstract fun bindCourseContinueViewContainer(@ProfileScope viewContainer: PresenterViewContainer<ProfileCoursesView>): ViewContainer<out CourseContinueView>
 
     @Module
     companion object {
@@ -92,11 +78,5 @@ abstract class ProfileModule {
         @ProfileScope
         fun providesUserObservable(subject: BehaviorSubject<ProfileData>, @BackgroundScheduler scheduler: Scheduler): Observable<ProfileData> =
             subject.observeOn(scheduler)
-
-        @Provides
-        @JvmStatic
-        @ProfileScope
-        fun provideViewContainer(): PresenterViewContainer<ProfileCoursesView> =
-            DefaultPresenterViewContainer()
     }
 }
